Adding featured photo
Photo bloggers can add a featured photo to a post. This featured photo
is shown on the summary page of their blog. Additionally, a short text
can be given which is displayed beneath the featured photo on the
blog's summary page. To use this feature, add variables
`featured_image` and `featured_text` to the front matter of you post's
markdown.